refactor: move the upload component and tweak styles
